Move InteractionHandler
Use existing RegistryLookupService in getBlockInteractionTarget
