+++ OptiGUI/src/main/java/opekope2/optigui/mixin/MinecraftClientMixin2.java
@@ -4,7 +4,7 @@ import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.gui.screen.Screen;
 import net.minecraft.client.network.ClientPlayerEntity;
 import net.minecraft.world.World;
-import opekope2.optigui.internal.InteractionHandler;
+import opekope2.optigui.internal.interaction.InteractionHandler;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
